Added the status of the order in initiatePayment

+++ src/main/java/com/example/demo/dto/BookRequest.java
@@ -1,7 +1,9 @@
 package com.example.demo.dto;
 
+import java.math.BigDecimal;
+
 public class BookRequest {
   public String title;
   public String author;
-  public int price;
+  public BigDecimal price;
 }
